adding new dataservice class
