+++ web/src/contexts/SysmonContext.tsx
@@ -17,7 +17,7 @@
 'use client';
 
 import React, { createContext, useContext, useEffect, useState, useCallback } from 'react';
-import { sysmonService, SysmonAgentSummary } from '@/services/sysmonService';
+import { dataService, SysmonAgentSummary } from '@/services/dataService';
 import { useAuth } from '@/components/AuthProvider';
 
 interface SysmonContextType {
@@ -38,7 +38,7 @@ export const SysmonProvider: React.FC<{ children: React.ReactNode }> = ({ childr
     const fetchData = useCallback(async () => {
         try {
             setError(null);
-            const sysmonData = await sysmonService.getSysmonData(token ?? undefined);
+            const sysmonData = await dataService.getSysmonData(token ?? undefined);
             setData(sysmonData);
         } catch (err) {
             console.error('Failed to fetch sysmon data:', err);
@@ -60,7 +60,7 @@ export const SysmonProvider: React.FC<{ children: React.ReactNode }> = ({ childr
         const interval = setInterval(fetchData, 60000);
         
         // Subscribe to service updates
-        const unsubscribe = sysmonService.subscribe(() => {
+        const unsubscribe = dataService.subscribeSysmon(() => {
             fetchData();
         });
